Refactor: relocate useScrollLock to _internal/overlay/
Same as the useFocusTrap relocation — prep for Drawer to share. No
behavior change.
